+++ Sources/ParseSwift/Parse.swift
@@ -52,12 +52,12 @@ public struct ParseConfiguration {
     public internal(set) var cacheDiskCapacity = 10_000_000
 
     /// If your app previously used the iOS Objective-C SDK, setting this value
-    /// to `true` will attempt to migrate relevant data stored in the Keychain to
-    /// ParseSwift. Defaults to `false`.
+    /// to **true** will attempt to migrate relevant data stored in the Keychain to
+    /// ParseSwift. Defaults to **false**.
     public internal(set) var isMigratingFromObjcSDK: Bool = false
 
     /// Deletes the Parse Keychain when the app is running for the first time.
-    /// Defaults to `false`.
+    /// Defaults to **false**.
     public internal(set) var isDeletingKeychainIfNeeded: Bool = false
 
     /// Maximum number of times to try to connect to Parse Server.
@@ -91,9 +91,9 @@ public struct ParseConfiguration {
      - parameter cacheMemoryCapacity: The memory capacity of the cache, in bytes. Defaults to 512KB.
      - parameter cacheDiskCapacity: The disk capacity of the cache, in bytes. Defaults to 10MB.
      - parameter migratingFromObjcSDK: If your app previously used the iOS Objective-C SDK, setting this value
-     to `true` will attempt to migrate relevant data stored in the Keychain to ParseSwift. Defaults to `false`.
+     to **true** will attempt to migrate relevant data stored in the Keychain to ParseSwift. Defaults to **false**.
      - parameter deletingKeychainIfNeeded: Deletes the Parse Keychain when the app is running for the first time.
-     Defaults to `false`.
+     Defaults to **false**.
      - parameter httpAdditionalHeaders: A dictionary of additional headers to send with requests. See Apple's
      [documentation](https://developer.apple.com/documentation/foundation/urlsessionconfiguration/1411532-httpadditionalheaders)
      for more info.
@@ -256,9 +256,9 @@ public struct ParseSwift {
      - parameter cacheMemoryCapacity: The memory capacity of the cache, in bytes. Defaults to 512KB.
      - parameter cacheDiskCapacity: The disk capacity of the cache, in bytes. Defaults to 10MB.
      - parameter migratingFromObjcSDK: If your app previously used the iOS Objective-C SDK, setting this value
-     to `true` will attempt to migrate relevant data stored in the Keychain to ParseSwift. Defaults to `false`.
+     to **true** will attempt to migrate relevant data stored in the Keychain to ParseSwift. Defaults to **false**.
      - parameter deletingKeychainIfNeeded: Deletes the Parse Keychain when the app is running for the first time.
-     Defaults to `false`.
+     Defaults to **false**.
      - parameter httpAdditionalHeaders: A dictionary of additional headers to send with requests. See Apple's
      [documentation](https://developer.apple.com/documentation/foundation/urlsessionconfiguration/1411532-httpadditionalheaders)
      for more info.