+++ Sources/ParseSwift/Types/ParseACL.swift
@@ -47,6 +47,17 @@ public struct ParseACL: ParseType,
     /// The default initializer.
     public init() { }
 
+    static func getRoleAccessName<R>(_ role: R) throws -> String where R: ParseRole {
+        guard let name = role.name else {
+            throw ParseError(code: .unknownError, message: "Name of ParseRole cannot be nil")
+        }
+        return getRoleAccessName(name)
+    }
+
+    static func getRoleAccessName(_ name: String) -> String {
+        return "role:\(name)"
+    }
+
     /**
      Controls whether the public is allowed to read this object.
     */
@@ -75,7 +86,7 @@ public struct ParseACL: ParseType,
      Returns true if a particular key has a specific access level.
      - parameter key: The key of the `ParseUser` or `ParseRole` for which to retrieve access.
      - parameter access: The type of access.
-     - returns: `true` if the `key` has *explicit* access, otherwise `false`.
+     - returns: **true** if the `key` has *explicit* access, otherwise **false**.
     */
     func get(_ key: String, access: Access) -> Bool {
         guard let acl = acl else { // no acl, all open!
@@ -87,11 +98,11 @@ public struct ParseACL: ParseType,
     // MARK: ParseUser
     /**
      Gets whether the given `objectId` is *explicitly* allowed to read this object.
-     Even if this returns `false`, the user may still be able to access it if `publicReadAccess` returns `true`
+     Even if this returns **false**, the user may still be able to access it if `publicReadAccess` returns **true**
      or if the user belongs to a role that has access.
 
      - parameter objectId: The `ParseUser.objectId` of the user for which to retrieve access.
-     - returns: `true` if the user with this `objectId` has *explicit* read access, otherwise `false`.
+     - returns: **true** if the user with this `objectId` has *explicit* read access, otherwise **false**.
     */
     public func getReadAccess(objectId: String) -> Bool {
         get(objectId, access: .read)
@@ -99,11 +110,11 @@ public struct ParseACL: ParseType,
 
     /**
      Gets whether the given `ParseUser` is *explicitly* allowed to read this object.
-     Even if this returns `false`, the user may still be able to access it if `publicReadAccess` returns `true`
+     Even if this returns **false**, the user may still be able to access it if `publicReadAccess` returns **true**
      or if the user belongs to a role that has access.
 
      - parameter user: The `ParseUser` for which to retrieve access.
-     - returns: `true` if the user with this `ParseUser` has *explicit* read access, otherwise `false`.
+     - returns: **true** if the user with this `ParseUser` has *explicit* read access, otherwise **false**.
     */
     public func getReadAccess<T>(user: T) -> Bool where T: ParseUser {
         if let objectId = user.objectId {
@@ -115,11 +126,11 @@ public struct ParseACL: ParseType,
 
     /**
      Gets whether the given `objectId` is *explicitly* allowed to write this object.
-     Even if this returns false, the user may still be able to write it if `publicWriteAccess` returns `true`
+     Even if this returns false, the user may still be able to write it if `publicWriteAccess` returns **true**
      or if the user belongs to a role that has access.
 
      - parameter objectId: The `ParseUser.objectId` of the user for which to retrieve access.
-     - returns: `true` if the user with this `ParseUser.objectId` has *explicit* write access, otherwise `false`.
+     - returns: **true** if the user with this `ParseUser.objectId` has *explicit* write access, otherwise **false**.
     */
     public func getWriteAccess(objectId: String) -> Bool {
         return get(objectId, access: .write)
@@ -127,11 +138,11 @@ public struct ParseACL: ParseType,
 
     /**
      Gets whether the given `ParseUser` is *explicitly* allowed to write this object.
-     Even if this returns false, the user may still be able to write it if `publicWriteAccess` returns `true`
+     Even if this returns false, the user may still be able to write it if `publicWriteAccess` returns **true**
      or if the user belongs to a role that has access.
 
      - parameter user: The `ParseUser` of the user for which to retrieve access.
-     - returns: `true` if the `ParseUser` has *explicit* write access, otherwise `false`.
+     - returns: **true** if the `ParseUser` has *explicit* write access, otherwise **false**.
     */
     public func getWriteAccess<T>(user: T) -> Bool where T: ParseUser {
         if let objectId = user.objectId {
@@ -189,48 +200,48 @@ public struct ParseACL: ParseType,
 
     /**
      Get whether users belonging to the role with the given name are allowed to read this object.
-     Even if this returns `false`, the role may still be able to read it if a parent role has read access.
+     Even if this returns **false**, the role may still be able to read it if a parent role has read access.
 
      - parameter roleName: The name of the role.
-     - returns: `true` if the role has read access, otherwise `false`.
+     - returns: **true** if the role has read access, otherwise **false**.
     */
     public func getReadAccess(roleName: String) -> Bool {
-        get(toRole(roleName: roleName), access: .read)
+        get(Self.getRoleAccessName(roleName), access: .read)
     }
 
     /**
      Get whether users belonging to the role are allowed to read this object.
-     Even if this returns `false`, the role may still be able to read it if a parent role has read access.
+     Even if this returns **false**, the role may still be able to read it if a parent role has read access.
 
      - parameter role: The `ParseRole` to get access for.
-     - returns: `true` if the `ParseRole` has read access, otherwise `false`.
+     - returns: **true** if the `ParseRole` has read access, otherwise **false**.
     */
     public func getReadAccess<T>(role: T) -> Bool where T: ParseRole {
         guard let name = role.name else { return false }
-        return get(toRole(roleName: name), access: .read)
+        return get(Self.getRoleAccessName(name), access: .read)
     }
 
     /**
      Get whether users belonging to the role with the given name are allowed to write this object.
-     Even if this returns `false`, the role may still be able to write it if a parent role has write access.
+     Even if this returns **false**, the role may still be able to write it if a parent role has write access.
 
      - parameter roleName: The name of the role.
-     - returns: `true` if the role has read access, otherwise `false`.
+     - returns: **true** if the role has read access, otherwise **false**.
     */
     public func getWriteAccess(roleName: String) -> Bool {
-        get(toRole(roleName: roleName), access: .write)
+        get(Self.getRoleAccessName(roleName), access: .write)
     }
 
     /**
      Get whether users belonging to the role are allowed to write this object.
-     Even if this returns `false`, the role may still be able to write it if a parent role has write access.
+     Even if this returns **false**, the role may still be able to write it if a parent role has write access.
 
      - parameter role: The `ParseRole` to get access for.
-     - returns: `true` if the role has read access, otherwise `false`.
+     - returns: **true** if the role has read access, otherwise **false**.
     */
     public func getWriteAccess<T>(role: T) -> Bool where T: ParseRole {
         guard let name = role.name else { return false }
-        return get(toRole(roleName: name), access: .write)
+        return get(Self.getRoleAccessName(name), access: .write)
     }
 
     /**
@@ -240,7 +251,7 @@ public struct ParseACL: ParseType,
      - parameter roleName: The name of the role.
     */
     public mutating func setReadAccess(roleName: String, value: Bool) {
-        set(toRole(roleName: roleName), access: .read, value: value)
+        set(Self.getRoleAccessName(roleName), access: .read, value: value)
     }
 
     /**
@@ -251,7 +262,7 @@ public struct ParseACL: ParseType,
     */
     public mutating func setReadAccess<T>(role: T, value: Bool) where T: ParseRole {
         guard let name = role.name else { return }
-        set(toRole(roleName: name), access: .read, value: value)
+        set(Self.getRoleAccessName(name), access: .read, value: value)
     }
 
     /**
@@ -261,7 +272,7 @@ public struct ParseACL: ParseType,
      - parameter roleName: The name of the role.
     */
     public mutating func setWriteAccess(roleName: String, value: Bool) {
-        set(toRole(roleName: roleName), access: .write, value: value)
+        set(Self.getRoleAccessName(roleName), access: .write, value: value)
     }
 
     /**
@@ -272,11 +283,7 @@ public struct ParseACL: ParseType,
     */
     public mutating func setWriteAccess<T>(role: T, value: Bool) where T: ParseRole {
         guard let name = role.name else { return }
-        set(toRole(roleName: name), access: .write, value: value)
-    }
-
-    private func toRole(roleName: String) -> String {
-        "role:\(roleName)"
+        set(Self.getRoleAccessName(name), access: .write, value: value)
     }
 
     private mutating func set(_ key: String, access: Access, value: Bool) {
@@ -365,10 +372,10 @@ extension ParseACL {
      This value will be copied and used as a template for the creation of new ACLs, so changes to the
      instance after this method has been called will not be reflected in new instance of `ParseObject`.
 
-     - parameter withAccessForCurrentUser: If `true`, the `ACL` that is applied to
+     - parameter withAccessForCurrentUser: If **true**, the `ACL` that is applied to
      newly-created instance of `ParseObject` will
      provide read and write access to the `ParseUser.+currentUser` at the time of creation.
-     - If `false`, the provided `acl` will be used without modification.
+     - If **false**, the provided `acl` will be used without modification.
      - If `acl` is `nil`, this value is ignored.
      
      - returns: Updated defaultACL